+++ lib/src/modules/engineer/screens/part_request_reasons_screen.g.dart
@@ -1,0 +1,22 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'part_request_reasons_screen.dart';
+
+// **************************************************************************
+// Generator: PageSchemaGen
+// **************************************************************************
+
+dynamic schema() => {
+      //#adminscreen
+      "type": "@@page",
+      "key": "TRC_part_request_reasons_screen",
+      "pageGroup": "Trc Part Request Reasons",
+      "title": "Part Request Reasons Screen",
+      "isActive": true,
+      "params": [
+        {"key": "prl", "value": null},
+        {"key": "rs", "value": null}
+      ]
+
+      //#adminscreen
+    };