+++ lib/src/modules/part_qc/models/qc_parts_list_response.g.dart
@@ -6,19 +6,25 @@ part of 'qc_parts_list_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-QcPartsListResponse _$QcPartsListResponseFromJson(Map<String, dynamic> json) => QcPartsListResponse(
+QcPartsListResponse _$QcPartsListResponseFromJson(Map<String, dynamic> json) =>
+    QcPartsListResponse(
       isSuccess: json['s'] as bool?,
       refId: json['r_id'] as String?,
-      dataList: (json['dt'] as List<dynamic>?)?.map((e) => QcPartListData.fromJson(e as Map<String, dynamic>)).toList(),
+      dataList: (json['dt'] as List<dynamic>?)
+          ?.map((e) => QcPartListData.fromJson(e as Map<String, dynamic>))
+          .toList(),
     );
 
-Map<String, dynamic> _$QcPartsListResponseToJson(QcPartsListResponse instance) => <String, dynamic>{
+Map<String, dynamic> _$QcPartsListResponseToJson(
+        QcPartsListResponse instance) =>
+    <String, dynamic>{
       'r_id': instance.refId,
       's': instance.isSuccess,
       'dt': instance.dataList,
     };
 
-QcPartListData _$QcPartListDataFromJson(Map<String, dynamic> json) => QcPartListData(
+QcPartListData _$QcPartListDataFromJson(Map<String, dynamic> json) =>
+    QcPartListData(
       prid: (json['prid'] as num?)?.toInt(),
       sku: json['sku'] as String?,
       partName: json['pn'] as String?,
@@ -31,7 +37,8 @@ QcPartListData _$QcPartListDataFromJson(Map<String, dynamic> json) => QcPartList
       partVariantName: json['pvn'] as String?,
     )..partColor = json['pc'] as String?;
 
-Map<String, dynamic> _$QcPartListDataToJson(QcPartListData instance) => <String, dynamic>{
+Map<String, dynamic> _$QcPartListDataToJson(QcPartListData instance) =>
+    <String, dynamic>{
       'prid': instance.prid,
       'sku': instance.sku,
       'pn': instance.partName,
@@ -40,7 +47,7 @@ Map<String, dynamic> _$QcPartListDataToJson(QcPartListData instance) => <String,
       'rqty': instance.requestedQuantity,
       'pbr': instance.partBarcode,
       'pc': instance.partColor,
+      'pvn': instance.partVariantName,
       'isDamaged': instance.isDamaged,
       'isBulk': instance.isBulk,
-      'pvn': instance.partVariantName,
     };